+++ module/disable-hooks.js
@@ -4,11 +4,11 @@ export class DisableHooks {
         Hooks.on("afterDamage", async actor => {
             await this.disableTalents(actor, ['damage'], []);
         });
-        
+
         Hooks.on("afterReduce", async actor => {
             await this.disableTalents(actor, ['reduce'], []);
         });
-        
+
         Hooks.on("afterRound", async actors => {
             for (let actor of actors)
                 await this.disableTalents(actor, ['damage','reduce', 'round'], ['round']);
@@ -18,7 +18,7 @@ export class DisableHooks {
             for (let actor of actors)
                 await this.disableTalents(actor, ['damage','reduce', 'round', 'battle'], ['round', 'battle']);
         });
-        
+
         Hooks.on("afterSession", async () => {
             for (let actor of game.actors) {
                 if (actor.type !== "enemy")
@@ -27,23 +27,27 @@ export class DisableHooks {
         });
 
     }
-    
+
     static async disableTalents(actor, active, used) {
         for (let item of actor.items) {
             let updates = {};
             if (item.data.data.active != undefined)
             if (active.findIndex(i => i == item.data.data.active.disable) != -1)
                 updates["data.active.state"] = false;
-            
+
             if (item.data.data.used != undefined)
             if (active.findIndex(i => i == item.data.data.used.disable) != -1)
                 updates["data.used.state"] = 0;
-                
+
             await item.update(updates);
         }
-        
+
+        let updates = {};
+        for (let [key, effect] of Object.entries(actor.data.data.attributes.effects)) {
+            if (active.findIndex(i => i == effect.disable) != -1)
+                updates[`data.attributes.effects.-=${key}`] = null;
+        }
+        await actor.update(updates);
     }
-    
-    
-    
+
 }