+++ module/sheet/item-sheet.js
@@ -93,7 +93,11 @@ export class KamigakariItemSheet extends ItemSheet {
 
       if (type == 'attributes') {
         const aNk = Object.keys(this.object.data.data.attributes).length + 1;
-        const attributes = `<input type="hidden" name="data.attributes.${aNk}.key" value="-" data-dType="String">`;
+        let attributes = "";
+        if (a.dataset.pos == "main")
+          attributes = `<input type="hidden" name="data.attributes.${aNk}.key" value="-" data-dType="String">`;
+        else
+          attributes = `<input type="hidden" name="data.effect.attributes.${aNk}.key" value="-" data-dType="String">`;
         newKey.innerHTML = attributes;
       } else if (type == 'additional') {
         const aNk = Object.keys(this.object.data.data.additional).length + 1;
@@ -132,7 +136,8 @@ export class KamigakariItemSheet extends ItemSheet {
     if (this.item.data.type == 'equipment') {
       formData = this.updateAdditional(formData);
       formData = this.updateMaterial(formData);
-    }
+    } else if (this.item.data.type == 'talent')
+      formData = this.updateEffectAttributes(formData);
 
     // Update the Item
     return this.object.update(formData);
@@ -160,6 +165,32 @@ export class KamigakariItemSheet extends ItemSheet {
       obj[e[0]] = e[1];
       return obj;
     }, {id: this.object.id, "data.attributes": attributes});
+
+    return formData;
+  }
+  
+  updateEffectAttributes(formData) {
+    // Handle the free-form attributes list
+    const formAttrs = expandObject(formData).data.effect.attributes || {};
+
+    const attributes = Object.values(formAttrs).reduce((obj, v) => {
+      let k = v["key"].trim();
+      if ( /[\s\.]/.test(k) )  return ui.notifications.error("Attribute keys may not contain spaces or periods");
+      delete v["key"];
+      obj[k] = v;
+      return obj;
+    }, {});
+
+    // Remove attributes which are no longer used
+    for ( let k of Object.keys(this.object.data.data.effect.attributes) ) {
+      if ( !attributes.hasOwnProperty(k) ) attributes[`-=${k}`] = null;
+    }
+
+    // Re-combine formData
+    formData = Object.entries(formData).filter(e => !e[0].startsWith("data.effect.attributes")).reduce((obj, e) => {
+      obj[e[0]] = e[1];
+      return obj;
+    }, {id: this.object.id, "data.effect.attributes": attributes});
 
     return formData;
   }